requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.12.0
vllm>=0.6.0
datasets>=3.0.0
tokenizers>=0.20.0
safetensors>=0.4.0
bitsandbytes>=0.44.0
flash-attn>=2.6.0
deepspeed>=0.15.0
scipy>=1.11.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
wandb>=0.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
einops>=0.8.0
timm>=1.0.0
sentencepiece>=0.2.0
qwen-vl-utils>=0.0.8
